+++ aiidalab_optimade/optimade_query.py
@@ -8,7 +8,7 @@ import ase
 from aiida.orm.nodes.data.structure import Site, Kind, StructureData
 
 from aiidalab_optimade.exceptions import InputError
-from aiidalab_optimade.helper_widgets import (
+from aiidalab_optimade.subwidgets import (
     ProvidersImplementations,
     StructureDropdown,
     FilterInputs,
@@ -29,22 +29,32 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
     NOTE: Only supports offset-pagination at the moment.
     """
 
-    structure = traitlets.Instance(ase.Atoms, allow_none=True)
+    structure = traitlets.Union(
+        [traitlets.Instance(ase.Atoms), traitlets.Instance(StructureData)],
+        allow_none=True,
+    )
     database = traitlets.Tuple(traitlets.Unicode(), traitlets.Dict(allow_none=True))
 
-    def __init__(self, debug: bool = False, result_limit: int = None, **kwargs):
+    def __init__(
+        self,
+        debug: bool = False,
+        embedded: bool = False,
+        result_limit: int = None,
+        **kwargs,
+    ):
         self.debug = debug
+        self.embedded = embedded
         self.page_limit = result_limit if result_limit else 10
         self.offset = 0
 
-        # self.header = ipw.HTML(
-        #     "<h4><strong>Search for a structure in an OPTiMaDe database</h4></strong>"
-        # )
-        self.base_url = ProvidersImplementations()
+        self.base_url = ProvidersImplementations(
+            include_summary=not self.embedded, debug=self.debug
+        )
         self.base_url.observe(self._on_database_select, names="database")
 
         self.filter_header = ipw.HTML("<br><h4>Apply filters</h4>")
         self.filters = FilterInputs()
+        self.filters.freeze()
         self.filters.on_submit(self.retrieve_data)
         self.query_button = ipw.Button(
             description="Search",
@@ -58,7 +68,7 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
         self.structures_header = ipw.HTML("<br><h4>Results</h4>")
         self.structure_drop = StructureDropdown(disabled=True)
         self.structure_drop.observe(self._on_structure_select, names="value")
-        self.structure_results_section = ipw.HTML("")
+        self.error_or_status_messages = ipw.HTML("")
 
         self.structure_page_chooser = ResultsPageChooser(self.page_limit)
         self.structure_page_chooser.observe(
@@ -67,14 +77,13 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
 
         super().__init__(
             children=[
-                # self.header,
                 self.base_url,
                 self.filter_header,
                 self.filters,
                 self.query_button,
                 self.structures_header,
                 self.structure_drop,
-                self.structure_results_section,
+                self.error_or_status_messages,
                 self.structure_page_chooser,
             ],
             layout=ipw.Layout(width="100%"),
@@ -87,9 +96,11 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
         if self.database[1] is None or self.database[1].get("base_url", None) is None:
             self.query_button.disabled = True
             self.query_button.tooltip = "Search - No database chosen"
+            self.filters.freeze()
         else:
             self.query_button.disabled = False
             self.query_button.tooltip = "Search"
+            self.filters.unfreeze()
         self.structure_drop.reset()
 
     def _on_structure_select(self, change):
@@ -100,7 +111,7 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
             with self.hold_trait_notifications():
                 self.structure_drop.index = 0
         else:
-            self.structure = chosen_structure["ase_atoms"]
+            self.structure = chosen_structure["structure"]
 
     def _get_more_results(self, change):
         """Query for more results according to page_offset"""
@@ -246,13 +257,14 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
 
         if "errors" in response:
             if "data" in response:
-                self.structure_results_section.value = (
-                    "Error(s) during querying, but "
+                self.error_or_status_messages.value = (
+                    '<font color="red">Error(s) during querying,</font> but '
                     f"<strong>{len(response['data'])}</strong> structures found."
                 )
             else:
-                self.structure_results_section.value = (
-                    "Error during querying, please try again later."
+                self.error_or_status_messages.value = (
+                    '<font color="red">Error during querying, '
+                    "please try again later.</font>"
                 )
             return True
 
@@ -268,7 +280,7 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
             if structure.has_vacancies or structure.is_alloy:
                 continue
 
-            formula = structure.get_formula()
+            formula = structure.get_formula(mode="hill_compact")
 
             optimade_id = entry["id"]
             entry_name = "{} (id={})".format(formula, optimade_id)
@@ -287,6 +299,10 @@ class OptimadeQueryWidget(ipw.VBox):  # pylint: disable=too-many-instance-attrib
             # Freeze and disable list of structures in dropdown widget
             # We don't want changes leading to weird things happening prior to the query ending
             self.freeze()
+
+            # Reset the error or status message
+            if self.error_or_status_messages.value:
+                self.error_or_status_messages.value = ""
 
             # Update button text and icon
             self.query_button.description = "Querying ... "
