+++ aiidalab_optimade/subwidgets/__init__.py
@@ -1,0 +1,13 @@
+# pylint: disable=undefined-variable
+from .filter_inputs import *
+from .output_summary import *
+from .provider_database import *
+from .results import *
+
+
+__all__ = (
+    filter_inputs.__all__  # noqa
+    + output_summary.__all__  # noqa
+    + provider_database.__all__  # noqa
+    + results.__all__  # noqa
+)
